+++ datanavigator/utils.py
@@ -14,7 +14,7 @@ from typing import List, Tuple, Union
 
 import cv2 as cv
 import numpy as np
-from decord import VideoReader, cpu
+from .video_reader import VideoReader, cpu
 from matplotlib import axes as maxes
 from matplotlib import pyplot as plt
 
@@ -319,7 +319,7 @@ class Video(VideoReader):
         Returns:
             np.ndarray: Grayscale frame.
         """
-        return cv.cvtColor(self[frame_num].asnumpy(), cv.COLOR_BGR2GRAY)
+        return cv.cvtColor(self[frame_num].asnumpy(), cv.COLOR_RGB2GRAY)
 
 
 def removeprefix(s: str, prefix: str) -> str: